+++ src/midje/error_handling/semi_sweet_errors.clj
@@ -8,6 +8,9 @@
     [midje.ideas.metaconstants :only [metaconstant-symbol?]]
     [midje.ideas.arrow-symbols :only [=contains=>]]))
 
+(defn- compiler-will-inline-fn? [fn]
+  (contains? (meta (resolve fn)) :inline))
+
 (defmethod validate "fake" [[_fake_ & fake-form :as form]]
   (let [funcall (first fake-form)]
     (cond (not (list? funcall))
@@ -15,6 +18,13 @@
             form
             "The left-hand-side of a prerequisite must look like a function call."
             (cl-format nil "`~S` doesn't." funcall))
+
+          (compiler-will-inline-fn? (first funcall))
+          (user-error-report-form
+           form
+           (cl-format nil "You cannot override the function `~S`: it is inlined by the Clojure compiler." (first funcall)))
+
+          
           :else
           fake-form)))